requirements.txt
torch
transformers
pandas
scikit-learn
numpy
PyYAML
fastapi
uvicorn
streamlit
lime
